requirements.txt
ExifRead==3.0.0
PyExifTool==0.5.6
